Ui improvements

See merge request alexxtheonly/empoweredpixels!74

+++ EmpoweredPixels/Models/Matches/MatchScoreBase.cs
@@ -18,7 +18,7 @@ namespace EmpoweredPixels.Models.Matches
 
     public int TotalDeaths { get; set; }
 
-    public float TotalDistanceTraveled { get; set; }
+    public double TotalDistanceTraveled { get; set; }
 
     public int TotalRegeneratedHealth { get; set; }
